requirements.txt
torch>=2.4.0
genesis-world
